[examples][BuddyLlama] Change some implementation

+++ examples/BuddyLlama/llama-main.cpp
@@ -7,10 +7,10 @@ using namespace buddy;
 using namespace std;
 
 extern "C" void _mlir_ciface_forward(MemRef<float, 3> *, MemRef<float, 1> *,
-                                     MemRef<long long, 2> *);
+                                     MemRef<size_t, 2> *);
 
 int main() {
-  // Guide the user to enter the VOCAB path
+  // Guide the user to enter the vocab path
   string vocabDir = "../../tests/Interface/core/vocab_llama.txt";
   // cout<<"please input vocab file path"<<endl;
   // getline(cin, vocabDir);
@@ -19,7 +19,7 @@ int main() {
   cout << "Please enter what you want to say to me" << endl;
   getline(cin, pureStr);
   clock_t buddyTokenizeStart = clock();
-  Text<long long, 2> pureStrContainer(pureStr);
+  Text<size_t, 2> pureStrContainer(pureStr);
   pureStrContainer.tokenizeLlama(vocabDir, 80);
   clock_t buddyTokenizeEnd = clock();
   double buddyTokenizeTime =
